requirements.txt
Django==2.0.7
django-ckeditor==5.6.1
django-js-asset==1.1.0
django-taggit==0.22.2
pkg-resources==0.0.0
pytz==2018.5
